refactor(pva-rs): own the EpicsValue<->PVA value-leaf mapping in one place
The DBF_CHAR-signedness bug (Q52) split across two files because the
serve path (native_source.rs) and the monitor-filter bridge
(server_native/tcp.rs) each hand-transcribed the same EpicsValue<->PvField
leaf correspondence: de500251 fixed the serve copy while the bridge copy
kept the old unsigned mapping (fixed separately in f9812674). Same
transcription, edited in one copy, missed in the other.

Collapse the correspondence to a single owner, `crate::leaf_convert`, so
that class of drift is unrepresentable — a signedness or type-code change
is edited once and every call site moves with it:

- epics_value_to_pv_leaf: EpicsValue -> PvField value leaf. Total over
  EpicsValue. Used by snapshot_to_pv_field (after its DBR_ENUM intercept)
  and the monitor-filter backward bridge (was epics_value_to_pv_field).
- epics_value_to_field_desc_leaf: the type-only projection, used by
  snapshot_to_field_desc, so served value and advertised type can't diverge.
- pv_leaf_to_epics_value: the partial inverse for the monitor-filter
  forward bridge (was pv_value_leaf_to_epics), preserving its fail-closed
  None for types the filter engine does not carry (ushort/uint gap intact).

Scope boundary, documented on the module: the shared owner covers only the
serve/monitor leaf direction. It deliberately does NOT absorb the inbound
PUT-decode path (native_source scalar_to_epics/typed_array_to_epics), whose
UByte->Char single-carrier fold is a different rule (put_pv re-coerces to
the record's FTVL, never serves a descriptor). Both review panels called
that boundary.

Tests: leaf_convert owns the mapping unit tests, incl. a forward/backward
round-trip that pins the two directions as genuine inverses and a test
pinning the ushort/uint forward gap so a later fill-in is deliberate. The
now-redundant direct-mapping test in tcp.rs is dropped; the end-to-end
filtered-DBF_CHAR-monitor test that exercises the wiring stays. No behavior
change: 1196 epics-pva-rs tests pass, clippy -D warnings clean.
